server/apps/users/views/token_obtain.py
from rest_framework_simplejwt.views import TokenObtainPairView

from ..serializers import RoleBasedTokenObtainPairSerializer


class RoleBasedTokenObtainPairView(TokenObtainPairView):
    serializer_class = RoleBasedTokenObtainPairSerializer
