server/pyproject.toml
[tool.poetry]
name = "Polls app"
version = "0.1.0"
description = "Polls app"
authors = ["Ilya Kormilkin <korid24.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "2.2.10"
psycopg2-binary = "2.8.6"
dj-database-url = "0.5.0"
gunicorn = "20.1.0"
ipython = "7.22.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.6.0"
django-filter = "2.4.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "3.2"
pep8-naming = "0.11.1"
flake8 = "3.9.0"
flake8-import-order = "0.18.1"
flake8-blind-except = "0.2.0"
flake8-builtins = "1.5.3"
flake8-broken-line = "0.3.0"
flake8-commas = "2.0.0"
flake8-eradicate = "1.0.0"
flake8-print = "4.0.0"
flake8-quotes = "3.2.0"
flake8-cognitive-complexity = "0.1.0"
flake8-bugbear = "21.4.3"
isort = "5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
